PythonServices/requirements.txt
pandas
scikit-learn
uvicorn
fastapi
joblib
mysql-connector-python
python-dotenv